+++ src/main/resources/templates/posts/show.html
@@ -13,6 +13,7 @@
   <link href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" rel="stylesheet">
   <script src="/js/newPost.js" defer></script>
   <script src="/js/main.js" defer></script>
+  <script src="/js/commentSend.js" defer></script>
 </head>
 <body>
   <div th:insert="fragments/nav :: navigation(${currentUserId})"></div>
